+++ coverage/cmdline.py
@@ -146,6 +146,13 @@ class Opts(object):
             "to be run as 'python -m' would run it."
         ),
     )
+    precision = optparse.make_option(
+        '', '--precision', action='store', metavar='N', type=int,
+        help=(
+            "Number of digits after the decimal point to display for "
+            "reported coverage percentages."
+        ),
+    )
     rcfile = optparse.make_option(
         '', '--rcfile', action='store',
         help=(
@@ -203,6 +210,7 @@ class CoverageOptionParser(optparse.OptionParser, object):
             omit=None,
             contexts=None,
             parallel_mode=None,
+            precision=None,
             pylib=None,
             rcfile=True,
             show_missing=None,
@@ -358,6 +366,7 @@ CMDS = {
             Opts.ignore_errors,
             Opts.include,
             Opts.omit,
+            Opts.precision,
             Opts.show_contexts,
             Opts.skip_covered,
             Opts.skip_empty,
@@ -395,6 +404,7 @@ CMDS = {
             Opts.ignore_errors,
             Opts.include,
             Opts.omit,
+            Opts.precision,
             Opts.show_missing,
             Opts.skip_covered,
             Opts.skip_empty,
@@ -583,6 +593,7 @@ class CoverageScript(object):
                 show_missing=options.show_missing,
                 skip_covered=options.skip_covered,
                 skip_empty=options.skip_empty,
+                precision=options.precision,
                 **report_args
                 )
         elif options.action == "annotate":
@@ -594,6 +605,7 @@ class CoverageScript(object):
                 skip_covered=options.skip_covered,
                 skip_empty=options.skip_empty,
                 show_contexts=options.show_contexts,
+                precision=options.precision,
                 **report_args
                 )
         elif options.action == "xml":